Inject a custom Future object in callf
Unfortunately concurrent.futures.ThreadPoolExecutor does not allow us to
specify a custom Future subclass to use. Instead, we have Agent.callf utilize
our own version, faws.futures.Future, which is *not* a subclass of the standard
Future, but takes an instance of it as a argument and delegates most item
accesses - except for a call to result().